+++ statique/src/main/java/ch/heigvd/igjt/statique/subcommands/SubCommandBuild.java
@@ -1,11 +1,11 @@
-package com.mycompany.app.subcommands;
+package ch.heigvd.igjt.subcommands.statique;
 
 import picocli.CommandLine;
 
 import java.util.concurrent.Callable;
 
-@CommandLine.Command(name = "serve")
-public class SubCommandServe implements Callable<Integer> {
+@CommandLine.Command(name = "build")
+public class SubCommandBuild implements Callable<Integer> {
     @Override
     public Integer call() throws Exception {
         return null;
