+++ statique/src/main/java/ch/heigvd/igjt/statique/subcommands/SubCommandServe.java
@@ -1,11 +1,11 @@
-package com.mycompany.app.subcommands;
+package ch.heigvd.igjt.subcommands;
 
 import picocli.CommandLine;
 
 import java.util.concurrent.Callable;
 
-@CommandLine.Command(name = "build")
-public class SubCommandBuild implements Callable<Integer> {
+@CommandLine.Command(name = "serve")
+public class SubCommandServe implements Callable<Integer> {
     @Override
     public Integer call() throws Exception {
         return null;
